Перенес две функции в файл util.js и удалил все комментарии из кода
