+++ js/util.js
@@ -1,4 +1,3 @@
-// Функция получения рандомного числа
 const getRandomNumber = function (minNum, maxNum, lengthNum = 0) {
   if (minNum < 0) {
     return 'Ошибка. Вы ввели отрицательное число. Число должно быть положительное.';
@@ -16,4 +15,20 @@ const getRandomNumber = function (minNum, maxNum, lengthNum = 0) {
 };
 getRandomNumber();
 
+const checkTextContent = (element) => {
+  if (element.textContent === '') {
+    element.remove();
+  }
+  return element;
+};
+
+const checkChild = (element) => {
+  if (element.children.length === 0) {
+    element.remove();
+  }
+  return element;
+};
+
 export {getRandomNumber};
+export {checkTextContent};
+export {checkChild};
